server_decomposed

+++ src/response/Response.hpp
@@ -1,13 +1,13 @@
 /* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   Response.hpp                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ymanfryd <marvin@42.fr>                    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/02/20 19:07:21 by ymanfryd          #+#    #+#             */
-/*   Updated: 2022/02/20 19:07:27 by ymanfryd         ###   ########.fr       */
-/*                                                                            */
+/*																			*/
+/*														:::	  ::::::::   */
+/*   Response.hpp									   :+:	  :+:	:+:   */
+/*													+:+ +:+		 +:+	 */
+/*   By: St.Krivtsov <1987stanislav@gmail.com>	  +#+  +:+	   +#+		*/
+/*												+#+#+#+#+#+   +#+		   */
+/*   Created: 2022/02/20 19:07:21 by ymanfryd		  #+#	#+#			 */
+/*   Updated: 2022/03/19 19:23:31 by St.Krivtsov	  ###   ########.fr	   */
+/*																			*/
 /* ************************************************************************** */
 
 #ifndef HTTP_WEBSERVER_RESPONSE_HPP
@@ -26,70 +26,70 @@ class Location;
 class Cgi;
 
 class Response {
-    private:
-            RequestParser                   RequestParser_;
-            ServerConfig                    ServerConfig_;
-            Location                        Location_;
-            std::string                     response_;
-            std::string                     requestRoute_;
-            std::string                     requestMethod_;
-            std::string                     requestBody_;
-            std::string                     responseHeaders_;
-            std::string                     responseContentType_;
-            std::string                     responseBody_;
-            int                             responseCode_;
-            int						        ClientMaxBodySize_;
-            size_t                          contentLength_;
-            std::vector<std::string>        requestPath_;
-            std::map<int, std::string>      responseCodes_;
-            std::map<std::string, Location> locations_;
-            std::set<std::string>           locationMethods_;
-            std::vector<std::string>        supportedMethods_;
-            std::vector<std::string>        locationIndex_;
-            std::string                     locationRoot_;
-            std::string                     locationRedirection_;
-            std::string                     requestedFile_;
-            std::map<int, std::string>		errorPages_;
-            bool                            cgiRequested_;
+	private:
+			RequestParser				   RequestParser_;
+			ServerConfig					ServerConfig_;
+			Location						Location_;
+			std::string					 response_;
+			std::string					 requestRoute_;
+			std::string					 requestMethod_;
+			std::string					 requestBody_;
+			std::string					 responseHeaders_;
+			std::string					 responseContentType_;
+			std::string					 responseBody_;
+			int							 responseCode_;
+			size_t						ClientMaxBodySize_;
+			size_t						  contentLength_;
+			std::vector<std::string>		requestPath_;
+			std::map<int, std::string>	  responseCodes_;
+			std::map<std::string, Location> locations_;
+			std::set<std::string>		   locationMethods_;
+			std::vector<std::string>		supportedMethods_;
+			std::vector<std::string>		locationIndex_;
+			std::string					 locationRoot_;
+			std::string					 locationRedirection_;
+			std::string					 requestedFile_;
+			std::map<int, std::string>		errorPages_;
+			bool							cgiRequested_;
 
-            void                            setResponse();
-            void                            setResponseCodes();
-            void                            setResponseHeaders();
-            void                            setContentType();
-            void                            setLocationMethods(const std::set<std::string>& locationMethods);
-            void                            setLocationIndex(const std::set<std::string>& locationIndex);
-            void                            setLocationRoot(const std::string& locationRoot);
-            void                            setLocationRedirection(const std::string& locationRedirection);
-            void                            setResponseCode(int code);
-            void                            setContentLength(size_t len);
-            void                            setResponseBody(const std::string& body);
+			void							setResponse();
+			void							setResponseCodes();
+			void							setResponseHeaders();
+			void							setContentType();
+			void							setLocationMethods(const std::set<std::string>& locationMethods);
+			void							setLocationIndex(const std::set<std::string>& locationIndex);
+			void							setLocationRoot(const std::string& locationRoot);
+			void							setLocationRedirection(const std::string& locationRedirection);
+			void							setResponseCode(int code);
+			void							setContentLength(size_t len);
+			void							setResponseBody(const std::string& body);
 
-            void                            createResponse();
-            void                            checkFileRequested();
-            void                            createAutoIndexPage(const char *dir);
-            void                            readLocationData();
-            void                            trimRequestPath();
-            int                             checkPathForLocation();
-            std::string                     findMaxPossibleLocation(const std::string& location);
-            std::string                     getScreen();
-            void                            savePostBody();
+			void							createResponse();
+			void							checkFileRequested();
+			void							createAutoIndexPage(const char *dir);
+			void							readLocationData();
+			void							trimRequestPath();
+			int							 checkPathForLocation();
+			std::string					 findMaxPossibleLocation(const std::string& location);
+			std::string					 getScreen();
+			void							savePostBody();
 
-            Response(): responseCode_(0), contentLength_(0) {};
-    public:
+			Response(): responseCode_(0), contentLength_(0) {};
+	public:
 
-            const std::string  &getResponse()          const;
-            const std::string  &getResponseBody()      const;
-            const std::string  &getResponseHeaders()   const;
-            const size_t       &getContentLength()     const;
-            const int          &getResponseCode()      const;
+			const std::string  &getResponse()		  const;
+			const std::string  &getResponseBody()	  const;
+			const std::string  &getResponseHeaders()   const;
+			const size_t	   &getContentLength()	 const;
+			const int		  &getResponseCode()	  const;
 
-            Response(RequestParser &request, ServerConfig &config);
-            Response(const Response &other);
-            Response &operator=(const Response &other);
-            ~Response() {}
+			Response(RequestParser &request, ServerConfig &config);
+			Response(const Response &other);
+			Response &operator=(const Response &other);
+			~Response() {}
 
-            static std::string readContent(const std::string &filename);
-            std::string  findFileWithExtension(std::string extension, std::string dir);
+			static std::string readContent(const std::string &filename);
+			std::string  findFileWithExtension(std::string extension, std::string dir);
 };
 
 #endif //HTTP_WEBSERVER_RESPONSE_HPP
